moved doubleArray to its own files for clarity
