+++ src/main/scala/nl/biopet/utils/Counts.scala
@@ -23,9 +23,8 @@ package nl.biopet.utils
 
 import java.io.{File, PrintWriter}
 
-import nl.biopet.utils.conversions.anyToJson
 import play.api.libs.json._
-import nl.biopet.utils.Counts.Implicits._
+
 import scala.collection.mutable
 
 /**
@@ -114,13 +113,13 @@ class Counts[T](c: Map[T, Long] = Map[T, Long]())(implicit ord: Ordering[T])
     * Converts to two IndexedSeqs, one containing values, and one containing counts.
     * @return Returns a doubleArray
     */
-  def toDoubleArray: Counts.DoubleArray[T] = {
+  def toDoubleArray: DoubleArray[T] = {
     val (keySeq, countSeq) =
       counts.foldLeft((IndexedSeq[T](), IndexedSeq[Long]())) {
         case ((keyList, countList), (key, count)) =>
           (keyList :+ key, countList :+ count)
       }
-    Counts.DoubleArray(keySeq, countSeq)
+    DoubleArray(keySeq, countSeq)
   }
 
 }
@@ -155,135 +154,17 @@ object Counts {
 
   /**
     * Converts a doublearray into a counts object
+    *
     * @param doubleArray the doubleArray
-    * @param ord Implicit ordering
+    * @param ord         Implicit ordering
     * @tparam T The type of the items in the values list.
     * @return A counts object.
     */
-  def fromDoubleArray[T](doubleArray: Counts.DoubleArray[T])(
+  def fromDoubleArray[T](doubleArray: DoubleArray[T])(
       implicit ord: Ordering[T]): Counts[T] = {
     new Counts[T](doubleArray.toMap)
   }
 
-  private case class Schema(map: Map[String, Long])
+  case class Schema(map: Map[String, Long])
 
-  object Implicits {
-
-    /**
-      * Method to write an indexedSeq to json.
-      *
-      * @tparam T IndexedSeq can be of any type.
-      * @return A Writes object
-      */
-    implicit def indexedSeqWrites[T]: Writes[IndexedSeq[T]] =
-      new Writes[IndexedSeq[T]] {
-        def writes(indexedSeq: IndexedSeq[T]): JsValue =
-          anyToJson(indexedSeq.toList.map(anyToJson))
-      }
-
-    /**
-      * Method to read an indexedSeq from json.
-      * @tparam T Can be of type int,long,double or string
-      * @return A writhe method.
-      */
-    implicit def indexedSeqReads[T]: Reads[IndexedSeq[T]] = {
-      new Reads[IndexedSeq[T]] {
-        def reads(json: JsValue): JsResult[IndexedSeq[T]] = {
-          // First evaluate if it can be parsed as an index holding a simple type
-          List(json.validate[List[Int]],
-               json.validate[List[Long]],
-               json.validate[List[Double]],
-               json.validate[List[String]])
-            .flatMap {
-              case JsSuccess(value: List[T], path) =>
-                Some(JsSuccess(value.toIndexedSeq, path))
-              case _ => None
-            }
-            .headOption
-            .getOrElse(JsError())
-        }
-      }
-    }
-
-    /**
-      * Read double array type.
-      * @tparam T The type of the items in the values list.
-      * @return a reads method.
-      */
-    implicit def doubleArrayReads[T]: Reads[Counts.DoubleArray[T]] =
-      new Reads[Counts.DoubleArray[T]] {
-        def reads(json: JsValue): JsResult[DoubleArray[T]] = {
-          json match {
-            case o: JsObject =>
-              val values: JsResult[IndexedSeq[T]] =
-                o.value("values").validate[IndexedSeq[T]]
-              val counts: JsResult[IndexedSeq[Long]] =
-                o.value("counts").validate[IndexedSeq[Long]]
-              JsSuccess(
-                DoubleArray(
-                  values.getOrElse(throw new IllegalStateException(
-                    "Values could not be parsed from json file.")),
-                  counts.getOrElse(throw new IllegalStateException(
-                    "Counts could not be parsed from json file."))
-                ))
-
-            case _ => throw new IllegalStateException("Not a object")
-          }
-        }
-      }
-
-    /**
-      * A method to write doublearrays
-      * @tparam T the type of the items in the value list.
-      * @return A json writes method.
-      */
-    implicit def doubleArrayWrites[T]: Writes[Counts.DoubleArray[T]] =
-      Json.writes[Counts.DoubleArray[T]]
-
-  }
-
-  /**
-    * A class that stores a T,Long dictionary as two sequences, that can be zipped.
-    * @param values An IndexedSeq of values
-    * @param counts An IndexedSeq of counts
-    * @tparam T A jsonifiable type
-    */
-  case class DoubleArray[T](values: IndexedSeq[T], counts: IndexedSeq[Long]) {
-    require(values.size == counts.size,
-            "Values and counts do not have the same length.")
-
-    def toMap: Map[T, Long] = this.values.zip(this.counts).toMap
-
-    def toJson: JsValue = Json.toJson(this)
-  }
-
-  object DoubleArray {
-
-    /**
-      * Convert a JsValue into a DoubleArray
-      * @param json a JsValue
-      * @tparam T The type of the items in the value list.
-      * @return a DoubleArray
-      */
-    def fromJson[T](json: JsValue): DoubleArray[T] = {
-      implicit def read: Reads[DoubleArray[T]] = Json.reads[DoubleArray[T]]
-      Json.reads[DoubleArray[T]].reads(json) match {
-        case x: JsSuccess[DoubleArray[T]] => x.value
-        case e: JsError                   => throw new IllegalStateException(e.toString)
-      }
-    }
-  }
-
-  /**
-    * Create a map[String,Long] using the schema
-    * @param json the jsvalue
-    * @return a map.
-    */
-  def mapFromJson(json: JsValue): Map[String, Long] = {
-    implicit val read: Reads[Schema] = Json.reads[Schema]
-    Json.reads[Schema].reads(json) match {
-      case x: JsSuccess[Schema] => x.value.map
-      case e: JsError           => throw new IllegalStateException(e.toString)
-    }
-  }
 }
